Convert some error into warnings durning sushi processing
Sentry by default creates an event when logger.error
is triggered. This can be helpful in some situations.
E.g. when an unhandled exception occurs during sushi processing.
=> we should fix the processing

On the other hand if SushiException is raised, should basically mean
that there is something wrong with the credentials or some unexpected
response (500, 404) was obtained from the server. But this doesn't
necessary mean that there is something wrong with our software.
Thus we don't want to create an issue in sentry for that.