Fix trigger_scheduler chaining
The problem is that during them manual based triggering
with trigger_scheduler(finish=True) a cron based triggering with
trigger_scheduler(finish=False) may occur. And if cron based triggering
is running in parallel with manual and cron based triggering holds
the lock, run_next() function return BUSY in manual and stops to
reschedule. So its chain of FetchIntentions is not finished.

This is not so crucial because these FetchIntetnions would be eventually
processed in cron tasks but with a longer delay (1/minute). But it
can annoy user a bit.

This patch makes sure that manual scheduling chain don't become broken.