Redo import task locking and suppress data already imported error
Remove cache_based_locks and use more granural select_for_update(skip_locked=True)
locking when importing batches of FetchAttempts

If a single FetchAttempt is imported, select_for_update(nowait=True) is
used so that if another task for the same FetchAttempts is
triggered, it raises DatabaseError. If the FetchAttempt is already
imported no error is propagated directly to FetchAttempt instance.
It didn't make much sense to mark these Attempts with import failed,
because the data were already imported...