fix: sometimes the number of broken credentials in the harvest reports email is lower than the one in UI
It was caused that `SushiFetchAttempt.when_processed` field was not set in some cases,
eventhough it should.

This patch tries to set the field properly and enforces using
`SushiCredentials.timestamp` instead. Timestamp is always present,
so we don't need to perform any db migration to fill in
`when_process` field.